Download device information from Senscape server using http get request.

Download device information from Senscape server using http get request.

Usage

get_senscape_devices(api_key)

Arguments

api_key

Senscape API key.

Value

A tibble.

Examples

my_devices = get_senscape_devices(api_key = Sys.getenv("SENSCAPE_API_KEY"))
my_devices
#> # A tibble: 10 × 6
#>    `_id`                    name               typ   dev_timestamp tags  created
#>    <chr>                    <chr>              <chr> <chr>         <lis> <chr>  
#>  1 5f1076ae98fda900151ff682 ASPB 3 (Pedralbes) Mosq… 2024-02-20T0… <chr> 2020-0…
#>  2 5f10767c98fda900151ff681 ASPB 2 (Horta)     Mosq… 2024-02-20T0… <chr> 2020-0…
#>  3 5f10762e98fda900151ff680 ASPB 1 (Sant Andr… Mosq… 2024-02-20T0… <chr> 2020-0…
#>  4 5f1076c998fda900151ff683 ASPB 4 (Lesseps)   Mosq… 2024-02-20T0… <chr> 2020-0…
#>  5 5f1076e798fda900151ff684 ASPB 5 (Zoo)       Mosq… 2024-02-19T2… <chr> 2020-0…
#>  6 60d1cf653f601d0014d4b581 IRTA - Baix Llobr… Mosq… 2024-02-10T0… <chr> 2021-0…
#>  7 5ffd9d37f0ca6600144ef426 Renegaedes 1       Mosq… 2023-12-19T1… <chr> 2021-0…
#>  8 61826eb99fc393001b27168e BIOVEC 1           Mosq… 2023-12-19T1… <chr> 2021-1…
#>  9 656df05cc51e070015751182 BPI 3              Mosq… 2023-12-12T0… <chr> 2023-1…
#> 10 656de773c51e070015751168 BPI 1              Mosq… 2023-12-12T0… <chr> 2023-1…