Download data from Senscape server using http get request.

Download data from Senscape server using http get request.

get_senscape_data(
  key_path,
  page_size = 10,
  start_datetime,
  end_datetime,
  deviceIds
)

Arguments

key_path

Path to Senscape API key.

page_size

The number items per page. Defaults to 10.

start_datetime

A datetime object that will be used as the lower bound to filter records.

end_datetime

A datetime object that will be used as the upper bound to filter records.

deviceIds

A character vector of all deviceIds from which data should be returned.

Value

A tibble.

Examples

new_smart_trap_data = get_senscape_data(key_path = "../auth/senscape_api_key.txt", start_datetime = lubridate::as_datetime("2023-03-08"), end_datetime = lubridate::as_datetime("2023-03-09"), deviceIds = c("5f1076c998fda900151ff683", "5f1076c998fda900151ff683", "5f10762e98fda900151ff680", "5f10767c98fda900151ff681", "5f1076ae98fda900151ff682"))
#> Gettingjj Senscape Data from 2023-03-08 to 2023-03-09
#>   |                                                                              |                                                                      |   0%
#> Error: '../auth/senscape_api_key.txt' does not exist in current working directory ('/opt/mosquitoR/docs/reference').
new_smart_trap_data
#> Error in eval(expr, envir, enclos): object 'new_smart_trap_data' not found